+++ src/main/java/com/thanhdat/quanlyhoctap/dto/request/EducationProgramCrudRequest.java
@@ -1,6 +1,5 @@
 package com.thanhdat.quanlyhoctap.dto.request;
 
-import com.thanhdat.quanlyhoctap.dto.response.EducationProgramCourseDto;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
@@ -13,6 +12,6 @@ import java.util.List;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class EducationProgramCrudRequest {
     int schoolYear;
-    Integer majorId;
+    Long majorId;
     List<EducationProgramCourseRequest> educationProgramCourses;
 }
