+++ src/main/java/com/thanhdat/quanlyhoctap/repository/ClassroomRepository.java
@@ -10,7 +10,7 @@ import java.util.Collection;
 import java.util.List;
 
 @Repository
-public interface ClassroomRepository extends JpaRepository<Classroom, Integer> {
+public interface ClassroomRepository extends JpaRepository<Classroom, Long> {
     @Query("SELECT c FROM Classroom c WHERE c.roomType = :roomType AND c.isAvailable = true AND c.id NOT IN :ids")
-    List<Classroom> findByRoomTypeAndIdNotIn(RoomType roomType, Collection<Integer> ids);
+    List<Classroom> findByRoomTypeAndIdNotIn(RoomType roomType, Collection<Long> ids);
 }
