+++ src/main/java/com/thanhdat/quanlyhoctap/repository/MajorRepository.java
@@ -8,11 +8,11 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface MajorRepository extends JpaRepository<Major, Integer> {
+public interface MajorRepository extends JpaRepository<Major, Long> {
     Optional<Major> findByName(String name);
 
     @Query("SELECT sc.major FROM StudentClass sc " +
             "LEFT JOIN sc.students st " +
             "WHERE st.id = :studentId")
-    Major getByStudentId(Integer studentId);
+    Major getByStudentId(Long studentId);
 }
