+++ src/main/java/com/thanhdat/quanlyhoctap/repository/SemesterRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface SemesterRepository extends JpaRepository<Semester, Integer> {
+public interface SemesterRepository extends JpaRepository<Semester, Long> {
     Semester findByYearAndSemester(Integer year, Integer semester);
 
     @Query("SELECT s FROM Semester s WHERE s.startDate <= :start AND s.startDate + ((s.durationWeeks * 7 - 1) day) >= :end")
@@ -19,14 +19,14 @@ public interface SemesterRepository extends JpaRepository<Semester, Integer> {
     @Query("SELECT epc.semester FROM EducationProgramCourse epc JOIN epc.educationProgram ep JOIN StudentClass sc ON " +
             "sc.major = ep.major AND sc.year = ep.schoolYear JOIN sc.students s WHERE s.id = :studentId " +
             "ORDER BY epc.semester.year DESC, epc.semester.semester DESC")
-    List<Semester> findByStudentId(Integer studentId);
+    List<Semester> findByStudentId(Long studentId);
 
     @Query("SELECT s FROM Semester s WHERE s.id = cast((SELECT st.value FROM Setting st WHERE st.key = 'SEMESTER_ID_FOR_REGISTER') as int)")
     Semester findForRegister();
 
     @Query("SELECT s FROM Semester s JOIN CourseClass cc ON cc.semester = s JOIN cc.teacher t WHERE t.id = :teacherId " +
             "ORDER BY s.year DESC, s.semester DESC")
-    List<Semester> findByTeacherId(Integer teacherId);
+    List<Semester> findByTeacherId(Long teacherId);
 
     List<Semester> findByLockTimeGreaterThan(LocalDateTime lockTime);
